Add logic to allow LSST to provide its own data to images
Now that the parameter namespace can be queried for flags set
by LSST without the code failing if they do not exist, it is
safe to put the image loading code in a block that will always
fire as long as LSST DM is not running the code. No changes
are necessary elsewhere to keep the original functionality.